refactor: move summary context to a separate file
