+++ src/store/SummaryProvider.js
@@ -1,13 +1,5 @@
-import { createContext } from "react";
-
-const SummaryContext = createContext({
-    selectedServers: [],
-    totalMemory: 0,
-    totalCPU: 0,
-    totalDisk: 0,
-    addServer: (item) => {},
-    removeServer: (id) => {}
-});
+import { useReducer } from "react";
+import SummaryContext from "./summary-context";
 
 const initialSummaryState = {
     selectedServers: [],
